+++ mbs/java/com/google/mbs/KeyBackupBucketProperties.java
@@ -39,6 +39,8 @@ public abstract class KeyBackupBucketProperties {
 
   public abstract String getTlogEntryPath();
 
+  public abstract String getCacheControl();
+
   @AutoValue.Builder
   public abstract static class Builder {
     public abstract KeyBackupBucketProperties.Builder setPublicBucketName(String value);
@@ -54,6 +56,8 @@ public abstract class KeyBackupBucketProperties {
     public abstract KeyBackupBucketProperties.Builder setAttestationDocPath(String value);
 
     public abstract KeyBackupBucketProperties.Builder setTlogEntryPath(String value);
+
+    public abstract KeyBackupBucketProperties.Builder setCacheControl(String value);
 
     public abstract KeyBackupBucketProperties build();
   }