+++ mbs/java/com/google/mbs/KeyBackupBucketPropertiesFactory.java
@@ -23,6 +23,7 @@ public class KeyBackupBucketPropertiesFactory {
   private static final String S3_CERT_NAME = "public/0/root_certificate.pem";
   private static final String S3_ATTESTATION_DOC_NAME = "public/0/attestation_doc.base64";
   private static final String S3_TLOG_ENTRY_NAME = "public/0/cert_tlog_entry.json";
+  private static final String DEFAULT_CACHE_CONTROL = "public, max-age=120";
 
   private final String publicBucketName;
   private final String privateBucketName;
@@ -51,6 +52,7 @@ public class KeyBackupBucketPropertiesFactory {
         .setCertPath(S3_CERT_NAME)
         .setAttestationDocPath(S3_ATTESTATION_DOC_NAME)
         .setTlogEntryPath(S3_TLOG_ENTRY_NAME)
+        .setCacheControl(DEFAULT_CACHE_CONTROL)
         .build();
   }
 }