+++ mbs/java/com/google/mbs/MbsCertificateFactory.java
@@ -19,6 +19,7 @@ package com.google.mbs;
 import java.math.BigInteger;
 import java.security.KeyPair;
 import java.security.KeyPairGenerator;
+import java.security.NoSuchAlgorithmException;
 import java.security.PrivateKey;
 import java.security.SecureRandom;
 import java.security.cert.CertificateException;
@@ -35,6 +36,7 @@ import org.bouncycastle.asn1.x509.KeyUsage;
 import org.bouncycastle.cert.CertIOException;
 import org.bouncycastle.cert.X509v3CertificateBuilder;
 import org.bouncycastle.cert.jcajce.JcaX509CertificateConverter;
+import org.bouncycastle.cert.jcajce.JcaX509ExtensionUtils;
 import org.bouncycastle.cert.jcajce.JcaX509v3CertificateBuilder;
 import org.bouncycastle.operator.ContentSigner;
 import org.bouncycastle.operator.OperatorCreationException;
@@ -120,6 +122,16 @@ public interface MbsCertificateFactory {
     certBuilder.addExtension(Extension.basicConstraints, true, new BasicConstraints(isCa));
 
     certBuilder.addExtension(Extension.keyUsage, true, new KeyUsage(keyUsageMask));
+
+    try {
+      JcaX509ExtensionUtils extUtils = new JcaX509ExtensionUtils();
+      certBuilder.addExtension(
+          Extension.subjectKeyIdentifier,
+          false,
+          extUtils.createSubjectKeyIdentifier(keyPair.getPublic()));
+    } catch (NoSuchAlgorithmException e) {
+      throw new CertificateException("Failed to add key identifiers", e);
+    }
 
     if (subjectAlternativeNames.isPresent()) {
       // RFC 5280 requires SAN to be critical if subject is empty.