+++ container-tools/CHANGELOG.md
@@ -2,6 +2,45 @@
 
 All notable changes to this project will be documented in this file. See [commit-and-tag-version](https://github.com/absolute-version/commit-and-tag-version) for commit guidelines.
 
+## 0.6.0 (2026-06-18)
+
+
+### Features
+
+* Support dynamic extra tags in AMI build packer template
+
+## 0.5.0 (2026-06-17)
+
+
+### Dependencies
+
+* **deps:** Update DevKit to release-3.9.0
+
+
+### Features
+
+* Tag AMIs with dynamic Git source repo URLs
+
+## 0.4.0 (2026-06-11)
+
+
+### Features
+
+* Defer CloudWatch Agent configuration to boot time
+
+## 0.3.0 (2026-06-10)
+
+
+### Dependencies
+
+* **deps:** Update DevKit to release-3.8.0
+
+
+### Features
+
+* Add metrics-exporter service
+* enable logback shutdown hook
+
 ## 0.2.0 (2026-05-25)