+++ mbs/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to this project will be documented in this file. See [commit-and-tag-version](https://github.com/absolute-version/commit-and-tag-version) for commit guidelines.
 
+## 0.4.0 (2026-06-10)
+
+
+### Dependencies
+
+* **deps:** Update DevKit to release-3.8.0
+
+
+### Features
+
+* Set cache-control attribute for files stored in s3 buckets
+* Set SubjectKeyIndentifier of generated cert
+
 ## 0.3.0 (2026-05-28)